Add a failing test for max-line-length config
As reported in #5, the current approach to using flake8 doesn't work for
all config files, because it skips some steps taken by the real flake8
CLI in parsing (for example) options such as 'max-line-length' into the
internal option name 'max_line_length'.

This commit adds a failing test for this case.