Use the flake8 CLI to run checks
This commit completely overhauls the check backend, replacing a
library-based implementation with one that shells out to the flake8 CLI.
This follows a move several other Hound linters have taken recently, and
fixes problems parsing some flake8 configuration directives.

As a result of this change, we can also upgrade to the latest version of
flake8, which has a) broken the library interface the previous code
relied upon and b) introduced support for customising the output format
of the linter, which makes parsing violations much easier.